pyproject.toml
[project]
name = "amplifier-app-server"
version = "0.1.0"
description = "Always-on Amplifier server runtime - hosts persistent AI sessions with HTTP/WebSocket API"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Brian Krabach" }
]
keywords = ["amplifier", "ai", "agent", "server", "api"]

dependencies = [
    # Amplifier core
    "amplifier-core @ git+https://github.com/microsoft/amplifier-core",
    "amplifier-foundation @ git+https://github.com/microsoft/amplifier-foundation",
    
    # Web framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    
    # Async support
    "anyio>=4.0",
    
    # Config and validation
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    
    # CLI
    "click>=8.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.25",  # For testing FastAPI
    "ruff>=0.1.0",
    "pyright>=1.1.350",
]

[project.scripts]
amplifier-server = "amplifier_server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/amplifier_server"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
